Refactor methods to be static
The `build` method is changed to be static, so a new Merkle Tree can be directly created from the provided elements. Before, the `build` function acted on an existing (empty) `MerkleTree`, which I consider to be unnecesary overhead.
Furthermore, other private methods that act as "utils" (hashing, tree level computation) have also been made static, as they do not act on the Merkle Tree data structure itself, only on the method inputs.